+++ boot/aether/src/boot/aether.clj
@@ -35,7 +35,7 @@
 
 (defn transfer-listener
   [{type :type meth :method {name :name repo :repository} :resource err :error :as info}]
-  (util/dbug "Aether: %s\n" (with-out-str (pprint/pprint info)))
+  (util/dbug* "Aether: %s\n" (with-out-str (pprint/pprint info)))
   (when (and (.endsWith name ".jar") (= type :started))
     (util/info "Retrieving %s from %s\n" (.getName (io/file name)) repo)))