+++ boot/core/src/boot/core.clj
@@ -130,14 +130,14 @@
 (defn- sync-user-dirs!
   []
   (util/with-semaphore-noblock sync-dirs-lock
-    (let [debug-mesg (delay (util/dbug "Syncing project dirs to temp dirs...\n"))]
+    (let [debug-mesg (delay (util/dbug* "Syncing project dirs to temp dirs...\n"))]
       (doseq [[k d] {:asset-paths    (user-asset-dirs)
                      :source-paths   (user-source-dirs)
                      :resource-paths (user-resource-dirs)
                      :checkout-paths @checkout-dirs}]
         @debug-mesg
         (patch! (first d) (get-env k) :ignore @bootignore))
-      (util/dbug "Sync complete.\n"))))
+      (util/dbug* "Sync complete.\n"))))
 
 (defn- set-fake-class-path!
   "Sets the fake.class.path system property to reflect all JAR files on the
@@ -243,7 +243,7 @@
               jar-dir   (.getParent (io/file jar-path))
               debounce  (or (:watcher-debounce env) 10)
               on-change (fn [_]
-                          (util/dbug "Refreshing checkout dependency %s...\n" (str p))
+                          (util/dbug* "Refreshing checkout dependency %s...\n" (str p))
                           (util/with-semaphore tempdirs-lock
                             (with-open [jarfs (fs/mkjarfs (io/file jar-path))]
                               (->> (patch! tmp [(fs/->path jarfs)] :state @tmp-state)
@@ -318,7 +318,7 @@
            warn# (delay (util/warn-deprecated @msg#))]
        (do (defn ~(with-meta was {:deprecated version}) [& args#]
              @warn#
-             (util/dbug (ex/format-exception (Exception. @msg#)))
+             (util/dbug* (ex/format-exception (Exception. @msg#)))
              (apply ~is args#))
            (alter-meta! #'~was assoc :doc @msg#)))))